[#57]refactor: hover transition 변경
