+++ client/src/components/productList/CategoryItem/index.tsx
