+++ client/src/styles/app.css
@@ -9,7 +9,7 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  min-height:1024px;
+  min-height: 1024px;
   align-items: center;
 }
